requirements.txt
torch
torchvision
segmentation-models-pytorch
opencv-python
numpy
Pillow
scikit-learn
streamlit
tqdm
matplotlib
shap